addons/ears/pyproject.toml
[project]
name = "whytcard-transcript"
version = "0.1.0"
description = "Audio transcription addon for WhytCard Hub"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "WhytCard Team"}
]
keywords = ["transcription", "whisper", "audio", "speech-to-text", "whytcard"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "openai-whisper>=20231117",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "httpx>=0.24.0",
    "aiofiles>=23.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "reportlab>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
whytcard-transcript = "src.cli:main"

[project.urls]
Homepage = "https://github.com/whytcard/WhytCard-Transcript"
Documentation = "https://whytcard.dev/docs/addons/transcript"
Repository = "https://github.com/whytcard/WhytCard-Transcript"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
